# Conflicts:
#	hawc/apps/animal/forms.py
#	hawc/apps/animal/templates/animal/animalgroup_copy_selector.html
#	hawc/apps/animal/templates/animal/animalgroup_detail.html
#	hawc/apps/animal/templates/animal/endpoint_copy_selector.html
#	hawc/apps/animal/templates/animal/experiment_copy_selector.html
#	hawc/apps/animal/templates/animal/experiment_detail.html
#	hawc/apps/animal/views.py
#	hawc/apps/common/forms.py
#	hawc/apps/epi/templates/epi/exposure_copy_selector.html
#	hawc/apps/epi/templates/epi/outcome_detail.html
#	hawc/apps/epi/templates/epi/studypopulation_detail.html
#	hawc/apps/epi/views.py
#	hawc/apps/epimeta/templates/epimeta/metaprotocol_detail.html
#	hawc/apps/lit/forms.py
#	hawc/apps/lit/managers.py
#	hawc/apps/study/templates/study/study_detail.html
#	hawc/apps/summary/forms.py
#	hawc/apps/summary/views.py
#	hawc/templates/hawc/_copy_as_new.html

+++ tests/hawc/tools/tables/test_set.py
@@ -21,7 +21,7 @@ class TestAttributeChoices:
         series = pd.Series({"foo": "foo", "bar": "bar", "free_html": "free_html"})
         cell = col.get_cell(series)
         assert cell.quill_text == "<p></p>"
-        series = pd.Series()
+        series = pd.Series(dtype=float)
         cell = col.get_cell(series)
         assert cell.quill_text == "<p></p>"
